Created Edit - Modified Add, Delete, and Register

+++ LAMPAPI/DeleteContact.php
@@ -1,6 +1,8 @@
 <?php
+
 	$inData = getRequestInfo();
 	
+  $UserID = $inData["UserID"];
 
 	$conn = new mysqli("localhost", "TheBeast", "WeLoveCOP4331", "COP4331");
 	if ($conn->connect_error) 
@@ -9,14 +11,26 @@
 	} 
 	else
 	{
-		$stmt = $conn->prepare("DELETE FROM Contacts WHERE UserID=? ");
-		$stmt->bind_param("s", $inData["UserID"]);
-		$stmt->execute();
-		$stmt->close();
-		$conn->close();
+		$stmt = $conn->prepare("SELECT * FROM Contacts WHERE UserID = ?");
+    $stmt->bind_param("s", $UserID);
+    $stmt->execute();
+    $result = $stmt->get_result();
+    if( $row = $result->fetch_assoc() )
+    {
+      $stmt = $conn->prepare("DELETE FROM Contacts WHERE UserID=? ");
+		  $stmt->bind_param("s", $UserID);
+		  $stmt->execute();
+		  $stmt->close();
+		  $conn->close();
 				
-    returnWithInfo($inData["UserID"]);
-		
+      returnWithInfo($UserID);
+    }
+    else
+    {
+      returnWithError("Contact with provided UserID does not exist.");
+      $stmt->close();
+      $conn->close();
+    }
 	}
 	
 	function getRequestInfo()
